requirements.txt
Django==5.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
